Revert "Add (de)serialization tests for `Message`." (commit 42aa888e117d87519501569d29101d85c56ac4b1). Commit was accidentally brought along with rebase of PR #1106.
