+++ Cargo.toml
@@ -152,8 +152,3 @@ extend-ignore-identifiers-re = [
     "RPL_[A-Z]+", # IRC Reply Numeric
     "TehPeGaSuS",
 ]
-
-[workspace.metadata.typos.files]
-extend-exclude = [
-  "data/tests/message/*.json",
-]
